handle consumer tag not found condition
This is likely in cases where an application restarts, and re-registers consumers with different id.
It is assumed that the application is not interested in past messages for older consumers.
If desired otherwise, applications should register consumers with fixed ids.

Also enabling locking on sendq after encountering some strange failures, possibly race conditions.