allow using tcp keepalives instead of heartbeats
AMQP allows heartbeats to be disabled completely, provided the server allows that.
But it is essential to have an alternative to detect network disconnections and stale
connections. In this PR, we switch on TCP keepalives on the connection (optional but
enabled by default) and allow heartbeats to be switched off. Heartbeats are still
enabled by default though.

The `connection` method has these new parameters to allow that:

```julia
connection(;
    heartbeat = true,
    keepalive = DEFAULT_KEEPALIVE_SECS,
)
```

- `heartbeat`: `true` to enable heartbeat, `false` to disable. Can also be set to a positive integer,
    in which case it is the heartbeat interval in seconds. Defaults to `true`. If `false`, ensure
    `keepalive` is enabled to detect dead connections. This parameter is negotiated with the server.
- `keepalive`: `true` to enable TCP keepalives, `false` to disable. Can also be set to a positive integer,
    in which case it is the keepalive interval in seconds. Defaults to `DEFAULT_KEEPALIVE_SECS`.